Move parameter wrapper classes to separate module
