+++ client/src/components/Workflow/Editor/Lint.vue
@@ -124,7 +124,7 @@
 <script>
 import Vue from "vue";
 import BootstrapVue from "bootstrap-vue";
-import { LegacyParameters } from "components/Workflow/Editor/modules/utilities";
+import { LegacyParameters } from "components/Workflow/Editor/modules/parameters";
 import LintSection from "components/Workflow/Editor/LintSection";
 import { getDisconnectedInputs, getInputsMissingMetadata, getWorkflowOutputs } from "./modules/utilities";
