+++ client/src/components/Workflow/Editor/modules/utilities.js
@@ -95,63 +95,6 @@ export function getStateUpgradeMessages(data) {
     return messages;
 }
 
-class LegacyParameterReference {
-    constructor(parameter, node) {
-        //this.node = node;
-        parameter.references.push(this);
-        this.nodeId = node.id;
-    }
-}
-
-class ToolInputLegacyParameterReference extends LegacyParameterReference {
-    constructor(parameter, node, tooInput) {
-        super(parameter, node);
-        this.toolInput = tooInput;
-    }
-}
-
-class PjaLegacyParameterReference extends LegacyParameterReference {
-    constructor(parameter, node, pja) {
-        super(parameter, node);
-        this.pja = pja;
-    }
-}
-
-class LegacyParameter {
-    constructor(name) {
-        this.name = name;
-        this.references = [];
-    }
-
-    canExtract() {
-        // Backend will indicate errors but would be a bit better to pre-check for them
-        // in a future iteration.
-        // return false if mixed input types or if say integers are used in PJA?
-        return true;
-    }
-}
-
-export class LegacyParameters {
-    constructor() {
-        this.parameters = [];
-    }
-
-    getParameter(name) {
-        for (const parameter of this.parameters) {
-            if (parameter.name == name) {
-                return parameter;
-            }
-        }
-        const legacyParameter = new LegacyParameter(name);
-        this.parameters.push(legacyParameter);
-        return legacyParameter;
-    }
-
-    getParameterFromMatch(match) {
-        return this.getParameter(match.substring(2, match.length - 1));
-    }
-}
-
 export function getDisconnectedInputs(nodes) {
     const inputs = [];
     Object.entries(nodes).forEach(([k, node]) => {
